update s3 location for nft artwork
