+++ packages/scoutgame/src/builderNfts/createBuilderNft.ts
@@ -7,22 +7,7 @@ import { currentSeason } from '../dates';
 import { builderApiClient } from './builderApiClient';
 import { builderContractAddress, builderNftChain } from './constants';
 import { generateNftImage } from './generateNftImage';
-
-function getS3ClientConfig() {
-  const config: Pick<S3ClientConfig, 'region' | 'credentials'> = {
-    region: process.env.S3_UPLOAD_REGION
-  };
-
-  if (process.env.S3_UPLOAD_KEY && process.env.S3_UPLOAD_SECRET) {
-    config.credentials = {
-      accessKeyId: process.env.S3_UPLOAD_KEY as string,
-      secretAccessKey: process.env.S3_UPLOAD_SECRET as string
-    };
-  }
-  return config;
-}
-
-const client = new S3Client(getS3ClientConfig());
+import { uploadArtwork } from './uploadArtwork';
 
 export async function createBuilderNft({
   avatar,
@@ -39,29 +24,12 @@ export async function createBuilderNft({
     args: { tokenId, amount: BigInt(1) }
   });
 
-  const imageBuffer = await generateNftImage({
+  const fileUrl = await uploadArtwork({
+    username,
+    season: currentSeason,
     avatar,
-    username
+    tokenId
   });
-
-  const s3Path = `nfts/seasons/${currentSeason}/beta/${tokenId}/artwork.png`;
-
-  const params: PutObjectCommandInput = {
-    ACL: 'public-read',
-    Bucket: process.env.SCOUTGAME_S3_BUCKET,
-    Key: s3Path,
-    Body: imageBuffer,
-    ContentType: 'image/png'
-  };
-
-  const s3Upload = new Upload({
-    client,
-    params
-  });
-
-  await s3Upload.done();
-
-  const fileUrl = `https://s3.amazonaws.com/${process.env.SCOUTGAME_S3_BUCKET}/${s3Path}`;
 
   const builderNft = await prisma.builderNft.create({
     data: {
